refactor common db models
