+++ Apps/Common/src/Database/Models/Route.cs
@@ -13,33 +13,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //-------------------------------------------------------------------------
-namespace HealthGateway.DIN.Models
+namespace HealthGateway.Common.Database.Models
 {
     using System;
     using System.ComponentModel.DataAnnotations;
-    using HealthGateway.Common.Database.Models;
 
-    public class Status : AuditableEntity
+    public class Route : AuditableEntity
     {
-        public Guid StatusId { get; set; }
+        public Guid RouteId { get; set; }
 
         [Required]
         public DrugProduct Drug { get; set; }
 
-        [MaxLength(1)]
-        public string CurrentStatusFlag { get; set; }
+        public int AdministrationCode { get; set; }
 
         [MaxLength(40)]
-        public string StatusDesc { get; set; }
+        public string Administration { get; set; }
 
         [MaxLength(80)]
-        public string StatusDescFrench { get; set; }
-
-        public DateTime? HistoryDate { get; set; }
-
-        [MaxLength(80)]
-        public string LotNumber { get; set; }
-
-        public DateTime? ExpirationDate { get; set; }
+        public string AdministrationFrench { get; set; }
     }
 }
