+++ Apps/Common/src/Database/Models/TherapeuticClass.cs
@@ -13,23 +13,34 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //-------------------------------------------------------------------------
-namespace HealthGateway.DIN.Models
+namespace HealthGateway.Common.Database.Models
 {
     using System;
     using System.ComponentModel.DataAnnotations;
-    using HealthGateway.Common.Database.Models;
 
-    public class Schedule : AuditableEntity
+    public class TherapeuticClass : AuditableEntity
     {
-        public Guid ScheduleId { get; set; }
+        public Guid TherapeuticClassId { get; set; }
 
         [Required]
         public DrugProduct Drug { get; set; }
 
-        [MaxLength(40)]
-        public string ScheduleDesc { get; set; }
+        [MaxLength(8)]
+        public string AtcNumber { get; set; }
+
+        [MaxLength(120)]
+        public string Atc { get; set; }
+
+        [MaxLength(240)]
+        public string AtcFrench { get; set; }
+
+        [MaxLength(20)]
+        public string AhfsNumber { get; set; }
 
         [MaxLength(80)]
-        public string ScheduleDescFrench { get; set; }
+        public string Ahfs { get; set; }
+
+        [MaxLength(160)]
+        public string AhfsFrench { get; set; }
     }
 }
