+++ Apps/Common/src/Database/Models/VeterinarySpecies.cs
@@ -13,38 +13,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //-------------------------------------------------------------------------
-namespace HealthGateway.DIN.Models
+namespace HealthGateway.Common.Database.Models
 {
     using System;
     using System.ComponentModel.DataAnnotations;
-    using HealthGateway.Common.Database.Models;
 
-    public class Packaging : AuditableEntity
+    public class VeterinarySpecies : AuditableEntity
     {
-        public Guid PackagingId { get; set; }
+        public Guid VeterinarySpeciesId { get; set; }
 
         [Required]
         public DrugProduct Drug { get; set; }
 
-        [MaxLength(12)]
-        public string UPC { get; set; }
+        [MaxLength(80)]
+        public string Species { get; set; }
 
-        [MaxLength(40)]
-        public string PackageType { get; set; }
+        [MaxLength(160)]
+        public string SpeciesFrench { get; set; }
 
         [MaxLength(80)]
-        public string PackageTypeFrench { get; set; }
-
-        [MaxLength(40)]
-        public string PackageSizeUnit { get; set; }
-
-        [MaxLength(80)]
-        public string PackageSizeUnitFrench { get; set; }
-
-        [MaxLength(5)]
-        public string PackageSize { get; set; }
-
-        [MaxLength(80)]
-        public string ProductInformation { get; set; }
+        public string SubSpecies { get; set; }
     }
 }
